movies_site/requirements.txt
Flask
requests
firebase-admin
python-dotenv

